Add support for leave time entry via ClickUp API
This commit introduces functionalities to create a time entry marking leave for a user on a given day using the ClickUp API. It involves extracting user information, their leave date, and tracking the time for the user on that date. Additional environment variables to support these functionalities have also been added to the .env.example file. Furthermore, aiohttp has been added as a new dependency.