Update application to support time tracking via AllHours
This commit updates the application to include support for AllHours in addition to ClickUp. Variables added to .env.example provide necessary configurations and the logic to interact with AllHours API has been included in the create_entry function. The push script has been removed and the README has been updated to reflect the changes.